* fixed wrong usage of size_t in time

+++ include/common/common-os/time/Const.h
@@ -34,29 +34,29 @@ namespace  common
     namespace time
     {
         /// Number of nanoseconds in a microsecond.
-        extern const size_t NSEC_PER_USEC;
+        extern const unsigned int NSEC_PER_USEC;
         /// Number of microseconds in a millisecond.
-        extern const size_t USEC_PER_MSEC;
+        extern const unsigned int USEC_PER_MSEC;
         /// Number of milliseconds in a second.
-        extern const size_t MSEC_PER_SEC;
+        extern const unsigned int MSEC_PER_SEC;
         /// Number of seconds in a minute.
-        extern const size_t SEC_PER_MIN;
+        extern const unsigned int SEC_PER_MIN;
         /// Number of minutes in an hour.
-        extern const size_t MIN_PER_HOUR;
+        extern const unsigned int MIN_PER_HOUR;
         /// Number of hours in a day.
-        extern const size_t HOUR_PER_DAY;
+        extern const unsigned int HOUR_PER_DAY;
         /// Number of days in a month.
-        extern const size_t DAY_PER_MONTH;
+        extern const unsigned int DAY_PER_MONTH;
         /// Number of months in a year.
-        extern const size_t MONTH_PER_YEAR;
+        extern const unsigned int MONTH_PER_YEAR;
 
         /// Number of nanoseconds in a WinTime-unit.
-        extern const size_t NSEC_PER_WINTIME;
+        extern const unsigned int NSEC_PER_WINTIME;
         /// Number of WinTime-units in a microsecond.
-        extern const size_t WINTIME_PER_USEC;
+        extern const unsigned int WINTIME_PER_USEC;
 
         /// Number of seconds between Unix Epoch (January 1, 1970) and Windows Epoch (January 1, 1601).
-        extern const size_t EPOCH_DIFF_SEC;
+        extern const unsigned long long EPOCH_DIFF_SEC;
     }
 }
